gfx/layers/apz/test/mochitest/helper_bug2041154-irregular-dynamic.html

<!DOCTYPE html>
<html>
<meta name="viewport" content="width=device-width; initial-scale=1.0">
<title>Test that preventDefault() in a dynamically-added touchmove listener suppresses scrolling on an irregular-area touchpoint</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<script src="/tests/SimpleTest/paint_listener.js"></script>
<script src="apz_test_utils.js"></script>
<script src="apz_test_native_event_utils.js"></script>
<style>
  #target {
    position: absolute;
    top: 0;
    left: 0;
    width: 100vw;
    height: 300px;
    background: green;
    clip-path: polygon(0% 0%, 100% 100%, 0% 100%);
  }
</style>
<div id="target"></div>
<div style="height: 500vh;"></div>
<script>
async function test() {
  // The initial hit-test result has eIrregularArea but no
  // eApzAwareListeners. A passive touchstart listener dynamically adds a
  // non-passive touchmove listener that calls preventDefault(). APZ must
  // discover the listener via the fast path and wait for a content
  // response on the first touchmove so preventDefault() is honored.
  const touchstartPromise = new Promise(resolve => {
    window.addEventListener("touchstart", () => {
      document.addEventListener("touchmove", e => {
        e.preventDefault();
      }, { passive: false, once: true });
      resolve();
    }, { passive: true, once: true });
  });

  await promiseApzFlushedRepaints();

  await synthesizeNativeTouch(window, 50, 200,
                              SpecialPowers.DOMWindowUtils.TOUCH_CONTACT);
  await touchstartPromise;

  // Though as of now fast-path notifiction explicitly schedules a paint,
  // it will not in future, so wait two requestAnimationFrames rather than
  // waiting for a paint.
  await promiseFrame();
  await promiseFrame();

  await synthesizeNativeTouch(window, 50, 150,
                              SpecialPowers.DOMWindowUtils.TOUCH_CONTACT);
  await synthesizeNativeTouch(window, 50, 100,
                              SpecialPowers.DOMWindowUtils.TOUCH_CONTACT);
  await synthesizeNativeTouch(window, 50, 100,
                              SpecialPowers.DOMWindowUtils.TOUCH_REMOVE);

  await promiseApzFlushedRepaints();

  is(window.scrollY, 0, "Should not have scrolled");
}

waitUntilApzStable()
.then(test)
.then(subtestDone, subtestFailed);
</script>
</html>
